requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
idna==3.2
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
pandas-datareader==0.10.0
Pillow==8.3.1
plotly==5.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.6
